Implementations generated by modules
